[Issue #25] Fixed a bug that caused the image to keep "growing".
The size of the cover was set to a size to place it within a bordered
area within the parent. But then the dimension of the cover were being
to set to the size of the image plus the border, which caused the parent
to then resize itself. This caused the image to resize, and feedback on
itself.