[BUG] Fixed the query params and user preferences fighting.
On the library page, changing the sort, rows, etc. would cause the
query parameters and the user preferences to bump into each other
changing the settings.

To avoid this, right now the are two gatekeeper flags to let each
of them only change settings once. But this STINKS. There has to be
a better way to handle this, but it's too late at night to think of
a fix right now...