Changed how the next/previous issue is fetched [#329]
When doing the search it first looks at the cover date and finds the comic
with the next higher/lower date. If two comics have the same cover date
then it compares the issue numbers to see which is higher/lower. Since no
two comics should ever have the same cover date and the same issue number,
this should be a sufficient check.