Added a new schema management system [#1707]
The new system handles both migrating the existing tables from
a v1.6 instance and also creating a new schema. As a result, the
old migrations are now all deleted to avoid packaging them with
future releases.

Removed the SmartListMatch and SmartReadingList types since
those tables are not going to be brought forward.