Fixed Spring Boot dependency for compatibility with JDK 21 [#1846]
Attempting to compile this under JDK 21 resulted in the following error:
Class com.sun.tools.javac.tree.JCTree$JCImport does not have member field 'com.sun.tools.javac.tree.JCTree qualid'

This error is in the Lombok library and was resolved in v1.18.30:
https://github.com/projectlombok/lombok/issues/3393

This project uses the latest version of Lombok which is compatible
with JDK 21, however it also uses Spring Boot which has its own
dependency which is to a specific version of JDK.

This patch updates the version of Lombok specified to be the minimum
version supported by JDK 21.  The result compiles, but I have not done
significant testing of the result (yet) to ensure there are no latent
issues.

Related Stack Overflow discussion:
https://stackoverflow.com/questions/77171270/compilation-error-after-upgrading-to-jdk-21-nosuchfielderror-jcimport-does-n