Fixed scraping comics with a hypen in the issue number [#2206]
The problem was that that "-" was getting filtered out of the
issue number due to how issue numbers are scrubbed.

This also exposed a bug with entering issue numbers with odd
characters or which are not simple ASCII. And the work around
for those is to have the admin find the reference id manually,
enter it, and then scrape using that id rather than searching.